healthapp/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
Flask==3.0.1
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
Werkzeug==3.0.1
zipp==3.17.0

